msm: kgsl: Optimize the function to search for free address range
kgsl_get_unmapped_area looks for a free address range that does
not collide in gpu map as well as the system map for the process.
Optimize this function by directly calling functions called within
get_unmappped_area. This also gives a better control of looking for
address ranges that have higher chance of success.

CRs-fixed: 562285
Change-Id: I530670b1753464f435d0c697a0ddaa8675a4e01d
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Signed-off-by: Eliminater74 <eliminater74@gmail.com>